+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * Bump standard to 13.1.0
 * Add scroll percent to context
 * Add custom state to context
+* Add `prng` function to utils
+* Add `choose` function to utils
 
 ## 4.0.0 (2019-05-10)